Add support to decl/init globals of primitive types
This adds another phase to the lowering to handle some
instructions in the top_inst_blck, as of now, only a
handful of them:

- VarStorage
- BindName
- Assign

To support the declaration/initialization of global variables of
primitive types (integer, float, bool).

The instructions that are not supported but still present in the top_inst_blck
will now not throw an error but instead they will be silently ignored.